Create list_Insertion_Sorted_Manner
Q. Write ADT code for an insert function for the list ADT in a way that the values are inserted in the sorted position 
in the list. For example, after inserting the values 4, 1, 5, 3, 2 in the list, the list should look like as shown in the 
following figure. The code should be generic and should also work for any other values.